backend/requirements.txt
Django==3.1.7
django-credit-cards==0.4.1
fontawesome-free==5.15.2
psycopg2-binary==2.8.6